Fixed mistake with merge
