+++ src/Chirp.Infrastructure/Services/CheepService/CheepService.cs
@@ -64,6 +64,11 @@ public class CheepService(ICheepRepository cheepRepository, IAuthorRepository au
         return await cheepRepository.DeleteCheepAsync((await cheepRepository.GetCheepByNotIDAsync(cheep.Author, cheep.Message, cheep.Timestamp)).Id);
     }
 
+    public async Task UpdateCheep(CheepViewModel cheep, string originalCheepMessage)
+    {
+        await cheepRepository.UpdateCheepAsync(CheepViewModelToCheepDTO(cheep), originalCheepMessage);
+    }
+
     public static CheepViewModel CheepDTOToCheepViewModel(CheepDTO cheepDTO)
     {
         return new(cheepDTO.Name, cheepDTO.Message, cheepDTO.TimeStamp);
