+++ src/Chirp.Razor/Pages/UserTimeline.cshtml
@@ -8,165 +8,5 @@
 
 <div>
     <h2> @routeName's Timeline </h2>
-    @if (User.Identity.IsAuthenticated)
-	{
-		<div class="cheepbox" style="min-height: 60px">
-			<h3>What's on your mind @(User.Identity.Name)?</h3>
-			<form method="post">
-				<input id="cheepText" style="float: left" type="text" asp-for="Text">
-				<p id="charCount" style="float: left">160 characters remaining</p>
-				<input id="btnShare" type="submit" value="Share" disabled>
-			</form>
-		</div>
-		<script>
-			document.getElementById('cheepText').addEventListener('input', function () {
-				const maxLength = 160;
-				const remaining = maxLength - this.value.length;
-				document.getElementById('charCount').innerText = remaining + ' characters remaining';
-
-				// https://stackoverflow.com/a/13831737
-				if (remaining == 160) document.getElementById("btnShare").disabled = true;
-				else document.getElementById("btnShare").disabled = false;
-			});
-		</script>
-	}
-	@if (Model.Cheeps.Any())
-    {
-        <ul id="messagelist" class="cheeps">
-            <script>
-			@Model.Incrementer;
-			var size = @Model.Cheeps.Count()
-			var editButtons = []
-			var endButtons = []
-			var paragraphs = []
-			var editedCheepMessageInputs = []
-            </script>
-            
-            @foreach (var cheep in Model.Cheeps)
-            {
-                <li>
-                    <p style="word-wrap:break-word;">
-
-                        @if (User.Identity.IsAuthenticated && !User.Identity.Name!.Equals(cheep.Author))
-                        {
-                            @if (await Model.AFollowsBAsync(User.Identity.Name, cheep.Author)) Model.FollowOrUnfollow = "Unfollow";
-                            else Model.FollowOrUnfollow = "Follow";
-                            <form method="post">
-                                <strong>
-                                    <a href="/@cheep.Author">@cheep.Author</a>
-                                </strong>
-                                <input type="hidden" name="cheepAuthor" value="@cheep.Author"/>
-                                <button type="submit" asp-page-handler="ChangeFollowStatus">@Model.FollowOrUnfollow</button>
-                            </form>
-                        }
-                        else
-                        {
-                            <strong>
-                                <a href="/@cheep.Author">@cheep.Author</a>
-                            </strong>
-                        }
-						<br />
-						@if (User.Identity.IsAuthenticated && User.Identity.Name!.Equals(cheep.Author))
-						{
-                            // CSS from: https://stackoverflow.com/questions/39972794/css-to-make-input-and-textarea-looks-like-simple-html-text - Atal Kishore
-                            <textarea id="edit_@Model.Incrementer" maxlength="160" style="background: rgba(0, 0, 0, 0); width: 100%; border: none; outline: 0; cursor: text; resize: none;">@cheep.Message</textarea>
-						}
-						else
-						{
-						<p>
-							@cheep.Message
-						</p>
-						}
-						<br />
-						<small>&mdash; @cheep.Timestamp</small>
-						@if (User.Identity.IsAuthenticated && User.Identity.Name!.Equals(cheep.Author))
-						{
-							<form	method="post">
-								<input type="hidden" name="cheepAuthor" value="@cheep.Author" />
-								<input type="hidden" name="cheepMessage" value="@cheep.Message" />
-								<input type="hidden" name="cheepTimestamp" value="@cheep.Timestamp" />
-								<input type="hidden" id="editedCheepMessage_@Model.Incrementer" name="newCheepMessage" value="" />
-								<button type="submit" asp-page-handler="DeleteCheep">Delete</button>
-								<button type="button" id="edit-button_@Model.Incrementer">Edit</button>
-								<button type="submit" id="end-editing_@Model.Incrementer" asp-page-handler="EditCheep">Done</button>
-							</form>
-
-							<script>
-								editButtons.push(document.getElementById("edit-button_@Model.Incrementer"))
-								endButtons.push(document.getElementById("end-editing_@Model.Incrementer"))
-								paragraphs.push(document.getElementById("edit_@Model.Incrementer"))
-								editedCheepMessageInputs.push(document.getElementById("editedCheepMessage_@Model.Incrementer"))
-								endButtons[@Model.Incrementer].disabled = true
-								endButtons[@Model.Incrementer].style.visibility = "hidden"
-
-								editButtons[@Model.Incrementer].addEventListener("click", function () {
-									paragraphs[@Model.Incrementer].contentEditable = true;
-									paragraphs[@Model.Incrementer].style.backgroundColor = "#dddbdb";
-									endButtons[@Model.Incrementer].disabled = false
-									endButtons[@Model.Incrementer].style.visibility = "visible"
-								});
-
-								endButtons[@Model.Incrementer].addEventListener("click", function () {
-									paragraphs[@Model.Incrementer].contentEditable = false;
-									paragraphs[@Model.Incrementer].style.backgroundColor = "#f0faf9";
-									editedCheepMessageInputs[@Model.Incrementer].value = paragraphs[@Model.Incrementer].value.trim();
-									endButtons[@Model.Incrementer].style.visibility = "hidden"
-								});
-								@Model.Increment()
-							</script>
-                        }
-                    </p>
-                </li>
-            }
-        </ul>
-
-        <nav aria-label="Page navigation">
-            <ul style="list-style-type: none; padding: 0; display: flex; justify-content: center;">
-                @if (Model.CurrentPage > 1)
-                {
-                    <li style="margin: 0 5px;">
-                        <a href="/@Model.RouteData.Values["author"]?page=@(Model.CurrentPage - 1)" aria-label="Previous" style="padding: 0.5em 1em; border: 1px solid #ddd; text-decoration: none;">&laquo;</a>
-                    </li>
-                }
-
-                @if (Model.CurrentPage > 3)
-                {
-                    <li style="margin: 0 5px;">
-                        <a href="/@Model.RouteData.Values["author"]?page=1" style="padding: 0.5em 1em; border: 1px solid #ddd; text-decoration: none;">1</a>
-                    </li>
-                    <li style="margin: 0 5px;">
-                        <span style="padding: 0.5em 1em; border: 1px solid #ddd;">...</span>
-                    </li>
-                }
-
-                @for (var i = Math.Max(1, Model.CurrentPage - 2); i <= Math.Min(Model.TotalPages, Model.CurrentPage + 2); i++)
-                {
-                    <li style="margin: 0 5px;" class="@(Model.CurrentPage == i ? "active" : "")">
-                        <a href="/@Model.RouteData.Values["author"]?page=@i" style="padding: 0.5em 1em; border: 1px solid #ddd; text-decoration: none;">@i</a>
-                    </li>
-                }
-
-                @if (Model.CurrentPage < Model.TotalPages - 2)
-                {
-                    <li style="margin: 0 5px;">
-                        <span style="padding: 0.5em 1em; border: 1px solid #ddd;">...</span>
-                    </li>
-                    <li style="margin: 0 5px;">
-                        <a href="/@Model.RouteData.Values["author"]?page=@Model.TotalPages" style="padding: 0.5em 1em; border: 1px solid #ddd; text-decoration: none;">@Model.TotalPages</a>
-                    </li>
-                }
-
-                @if (Model.CurrentPage < Model.TotalPages)
-                {
-                    <li style="margin: 0 5px;">
-                        <a href="/@Model.RouteData.Values["author"]?page=@(Model.CurrentPage + 1)" aria-label="Next" style="padding: 0.5em 1em; border: 1px solid #ddd; text-decoration: none;">&raquo;</a>
-                    </li>
-                }
-            </ul>
-        </nav>
-    }
-    else
-    {
-        <em>There are no cheeps so far.</em>
-    }
+    <partial name="Shared/_CheepTimeline" />
 </div>
